before dto added

+++ src/main/java/com/landproject/repository/InvestorRepository.java
@@ -3,5 +3,5 @@ package com.landproject.repository;
 import com.landproject.model.Investor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InvestorRepository extends JpaRepository<Investor, Integer> {
+public interface InvestorRepository extends JpaRepository<Investor, Long> {
 }
